COPL-888: optimize the number of calls to get the groups in which a register type is contained, need only to trigger the cache cleaning as the rest was already done earlier

+++ comu/com.energyict.mdc.masterdata/src/main/java/com/energyict/mdc/masterdata/impl/RegisterGroupImpl.java
@@ -1,5 +1,6 @@
 package com.energyict.mdc.masterdata.impl;
 
+import com.energyict.mdc.masterdata.MasterDataService;
 import com.energyict.mdc.masterdata.RegisterGroup;
 import com.energyict.mdc.masterdata.RegisterType;
 import com.energyict.mdc.masterdata.exceptions.MessageSeeds;
@@ -9,6 +10,8 @@ import com.elster.jupiter.domain.util.Save;
 import com.elster.jupiter.events.EventService;
 import com.elster.jupiter.nls.Thesaurus;
 import com.elster.jupiter.orm.DataModel;
+import com.elster.jupiter.orm.InvalidateCacheRequest;
+import com.elster.jupiter.pubsub.Publisher;
 import org.hibernate.validator.constraints.NotEmpty;
 
 import javax.inject.Inject;
@@ -22,7 +25,9 @@ import java.util.stream.Collectors;
 
 public class RegisterGroupImpl extends PersistentNamedObject<RegisterGroup> implements RegisterGroup {
 
+    private final Publisher publisher;
     private List<RegisterTypeInGroup> registerTypeInGroups = new ArrayList<>();
+    private ChangeNotifier changeNotifier = new NotNotifiedYet();
 
     @NotNull(groups = {Save.Create.class, Save.Update.class}, message = "{" + MessageSeeds.Keys.NAME_REQUIRED + "}")
     @NotEmpty(groups = {Save.Create.class, Save.Update.class}, message = "{" + MessageSeeds.Keys.NAME_REQUIRED + "}")
@@ -30,17 +35,18 @@ public class RegisterGroupImpl extends PersistentNamedObject<RegisterGroup> impl
     private String name;
 
     @Inject
-    public RegisterGroupImpl(DataModel dataModel, EventService eventService, Thesaurus thesaurus) {
+    public RegisterGroupImpl(DataModel dataModel, EventService eventService, Thesaurus thesaurus, Publisher publisher) {
         super(RegisterGroup.class, dataModel, eventService, thesaurus);
+        this.publisher = publisher;
+    }
+
+    static RegisterGroupImpl from (DataModel dataModel, String name) {
+        return dataModel.getInstance(RegisterGroupImpl.class).initialize(name);
     }
 
     RegisterGroupImpl initialize (String name) {
         this.setName(name);
         return this;
-    }
-
-    static RegisterGroupImpl from (DataModel dataModel, String name) {
-        return dataModel.getInstance(RegisterGroupImpl.class).initialize(name);
     }
 
     @Override
@@ -51,11 +57,14 @@ public class RegisterGroupImpl extends PersistentNamedObject<RegisterGroup> impl
     @Override
     protected void doDelete() {
         this.removeRegisterTypes();
-        this.getDataMapper().remove(this);
+        if (this.getId() > 0) {
+            this.getDataMapper().remove(this);
+            this.changeNotifier.deleted();
+        }
     }
 
     private void removeRegisterTypes() {
-        this.registerTypeInGroups.clear();;
+        this.registerTypeInGroups.clear();
     }
 
     @Override
@@ -89,9 +98,7 @@ public class RegisterGroupImpl extends PersistentNamedObject<RegisterGroup> impl
     @Override
     public void addRegisterType(RegisterType registerType) {
         this.registerTypeInGroups.add(RegisterTypeInGroup.from(this.dataModel, this, registerType));
-        if (this.getId() > 0) {
-            this.dataModel.touch(this);
-        }
+        this.changeNotifier.added(registerType);
     }
 
     @Override
@@ -104,7 +111,7 @@ public class RegisterGroupImpl extends PersistentNamedObject<RegisterGroup> impl
             }
         }
         this.checkAtLeastOneRegisterType();
-        this.dataModel.touch(this);
+        this.changeNotifier.removed(registerType);
     }
 
     @Override
@@ -112,7 +119,7 @@ public class RegisterGroupImpl extends PersistentNamedObject<RegisterGroup> impl
         this.removeObsoleteRegisterTypes(registerTypes);
         this.addNewRegisterTypes(registerTypes);
         this.checkAtLeastOneRegisterType();
-        this.dataModel.touch(this);
+        this.changeNotifier.updated();
     }
 
     private void checkAtLeastOneRegisterType() {
@@ -159,4 +166,104 @@ public class RegisterGroupImpl extends PersistentNamedObject<RegisterGroup> impl
         this.name = name;
     }
 
+    @Override
+    protected void postNew() {
+        super.postNew();
+        this.resetChangeNotifier();
+    }
+
+    @Override
+    protected void post() {
+        super.post();
+        this.resetChangeNotifier();
+    }
+
+    private void cleanRegisterTypeCache() {
+        this.publisher.publish(new InvalidateCacheRequest(MasterDataService.COMPONENTNAME, TableSpecs.MDS_MEASUREMENTTYPE.name()));
+    }
+
+    private void touchIfPersistent() {
+        if (this.getId() > 0) {
+            this.dataModel.touch(this);
+        }
+    }
+
+    private void resetChangeNotifier() {
+        this.changeNotifier = new NotNotifiedYet();
+    }
+
+    private void alreadyNotified() {
+        this.changeNotifier = new AlreadyNotified();
+    }
+
+    /**
+     * Notifies others of changes that were applied
+     * to this RegisterGroup, including the OrmService
+     * that is notified that the version needs to be bumped.
+     */
+    private interface ChangeNotifier {
+
+        void added(RegisterType registerType);
+
+        void removed(RegisterType registerType);
+
+        void updated();
+
+        void deleted();
+
+    }
+
+    /**
+     * Provides an implementation for the ChangeNotifier interface
+     * that will notify the interested parties and then switch over
+     * to another component that will no longer notify interested
+     * parties as one notification is enough.
+     */
+    private class NotNotifiedYet implements ChangeNotifier {
+        @Override
+        public void added(RegisterType registerType) {
+            this.updated();
+        }
+
+        @Override
+        public void removed(RegisterType registerType) {
+            this.updated();
+        }
+
+        @Override
+        public void updated() {
+            cleanRegisterTypeCache();
+            touchIfPersistent();
+            alreadyNotified();
+        }
+
+        @Override
+        public void deleted() {
+            cleanRegisterTypeCache();
+        }
+    }
+
+    /**
+     * Provides an implementation for the ChangeNotifier interface
+     * that is used when the other components have already been
+     * notified. Sending out one notification is enough
+     * to avoid that we are sued for stalking ;-)
+     */
+    private class AlreadyNotified implements ChangeNotifier {
+        @Override
+        public void added(RegisterType registerType) {
+        }
+
+        @Override
+        public void removed(RegisterType registerType) {
+        }
+
+        @Override
+        public void updated() {
+        }
+
+        @Override
+        public void deleted() {
+        }
+    }
 }
